requirements_tally.txt
accelerate==1.5.2
ace_tools==0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
azure-ai-inference==1.0.0b9
azure-core==1.34.0
azure-identity==1.23.0
beautifulsoup4==4.13.4
cffi==1.17.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.3
cycler==0.12.1
datasets==3.5.0
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
dpkt==1.9.8
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fonttools==4.56.0
frozenlist==1.5.0
googleapis-common-protos==1.70.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.30.1
importlib_metadata==8.6.1
instructor==1.8.3
ipykernel==6.29.5
ipython==8.37.0
isodate==0.7.2
jedi==0.19.2
jiter==0.8.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.1.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.9.2
-e git+https://github.com/lastmile-ai/mcp-agent.git@23fbe0997615589557ff79c7d0fa5d94616ce8a2#egg=mcp_agent
mcp-server-fetch==2025.4.7
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
nest-asyncio==1.6.0
numba==0.61.0
openai==1.83.0
opentelemetry-api==1.33.1
opentelemetry-distro==0.54b1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-anthropic==0.40.7
opentelemetry-instrumentation-openai==0.40.7
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-semantic-conventions-ai==0.4.9
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
pooch==1.8.2
prompt_toolkit==3.0.51
propcache==0.3.1
Protego==0.4.0
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
py3nvml==0.2.7
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyzmq==27.0.0
readabilipy==0.3.0
regex==2024.11.6
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torchaudio==2.2.0
tornado==6.5.1
traitlets==5.14.3
transformers==4.50.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.2
uvicorn==0.34.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0